refactor: split src/db/queries/ — relocate boundary helpers + remove server-action files
Auth-guard helpers (requireSession / requireUser / requireAdmin) and
the API-key request validator (validateApiKey) are HTTP-boundary
concerns, not database queries. They were sitting in src/db/queries/
behind a misleading directory name.

Relocated next to responses.mjs and methodNotAllowed.mjs:
  - src/db/queries/_authGuards.mjs    → src/shared/utils/api/authGuards.mjs
  - src/db/queries/validateApiKey.mjs → src/shared/utils/api/validateApiKey.mjs

Also removes (without replacement in this commit) the three
server-action files that move to feature directories in follow-up
commits. Their consumers still reference the old paths through HEAD~1;
the next two commits in this branch land the merged feature files and
update every consumer in lockstep. Reviewing as a series is cleaner
than reviewing each commit standalone.

Removed (relocated in commits 2 + 3):
  - src/db/queries/admin.mjs    (7 admin actions → features/admin/actions.mjs)
  - src/db/queries/api.mjs      (3 API key actions → features/account/actions.mjs)
  - src/db/queries/account.mjs  (2 user lifecycle actions → features/account/actions.mjs)

Updated importers in this commit (only the two outside the relocation set):
  - src/features/archives/reseedSeason.mjs    (requireAdmin)
  - src/app/api/h1/rebroadcast/route.js       (validateApiKey)

Plus two test files and the two docs MDX pages.
